Main now endlessly loops the Gameboy, and added Logging.
The Gameboy now ticks indefinitely (until the user quits the program
using CTRL+C or other means), but when a rendering target is added, the
quit button will be used to stop the program.

Logging was added as to not spam the console. Currently the apu, cpu and
ppu print "[x] tick!", but only when 'RUST_LOG=debug' is prepended to
running the program.